+++ agate/aggregations.py
@@ -21,7 +21,7 @@ import math
 
 from agate.data_types import Boolean, Date, DateTime, Number, Text
 from agate.exceptions import DataTypeError, UnsupportedAggregationError
-from agate.utils import Quantiles, max_precision, median
+from agate.utils import Quantiles, default, max_precision, median
 from agate.warns import warn_null_calculation
 
 
@@ -165,30 +165,20 @@ class All(Aggregation):
         return all(self._test(d) for d in data)
 
 
-class Length(Aggregation):
-    """
-    Count the total number of values in the column.
-
-    Equivalent to calling :func:`len` on a :class:`.Column`.
-    """
-    def get_aggregate_data_type(self, table):
-        return Number()
-
-    def run(self, table):
-        return len(table.rows)
-
-
 class Count(Aggregation):
     """
-    Count the number of times a specific value occurs in a column.
+    Count values. If no arguments are specified, this is simply a count of the
+    number of rows in the table. If only :code:`column_name` is specified, this
+    will count the number of non-null values in that column. If both
+    :code:`column_name` and :code:`value` are specified, then it will count
+    occurrences of a specific value in the specified column will be counted.
 
-    If you want to count the total number of rows in a column use
-    :class:`Length`.
-
+    :param column_name:
+        A column to count values in.
     :param value:
         Any value to be counted, including :code:`None`.
     """
-    def __init__(self, column_name, value):
+    def __init__(self, column_name=None, value=default):
         self._column_name = column_name
         self._value = value
 
@@ -196,7 +186,13 @@ class Count(Aggregation):
         return Number()
 
     def run(self, table):
-        return table.columns[self._column_name].values().count(self._value)
+        if self._column_name is not None:
+            if self._value is not default:
+                return table.columns[self._column_name].values().count(self._value)
+            else:
+                return len(table.columns[self._column_name].values_without_nulls())
+        else:
+            return len(table.rows)
 
 
 class Min(Aggregation):